app/requirements.txt
python-telegram-bot
requests
mysql-connector-python